Refactor process module for command processing and response generation
This commit introduces a refactored process module to enhance the
command processing and response generation functionality. The code has
been cleaned and restructured, improving readability and
maintainability. Additionally, the procces file has been renamed to
process for consistency. This refactor enables better integration with
Virgil API and other APIs, resulting in more accurate and timely
responses. The changes also include the addition of logging and event
handling functionalities.

No references to issues are made in this commit.
